refactor(setup): collapse the wizard to one route, drop the module singleton
The four step routes (/setup, /setup/access, /setup/activation,
/setup/practice) forced shared wizard state — permission probing,
practice success — into a module singleton (setup-state.svelte.ts),
because separate route components can't share $state.

Collapse to a single /setup route that carries the step in the URL as
?step=<id>. SvelteKit preserves the page component instance across
query-only navigation, so the shared state lives on that one instance
(permissions via a createSetupPermissions() factory, practice success as
local $state) while back/forward, deep links, and refresh still work.

Each step body becomes a child component (EngineStep/AccessStep/
ActivationStep/PracticeStep) receiving state by prop. Deletes the
layout, the singleton, the three step routes, and the path<->step
mapping (stepFromPath/pathFor) with its engine<->/setup special case.

Bonus: practice success now resets when the wizard is re-entered
(component state) instead of persisting for the app's lifetime.
